:rocket: Updated bottom navigation with new menus, colors and restructured code.
