+++ app/src/main/res/drawable/selector_navigation_color.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="true" android:color="@color/colorAccent" />
+    <item android:color="@android:color/darker_gray"  />
+</selector>
