+++ app/src/main/res/drawable/selector_navigation_icon_home.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/ic_home" android:state_checked="true"/>
+    <item android:drawable="@drawable/ic_outline_home" android:state_checked="false"/>
+</selector>
