+++ app/src/main/res/menu/navigation.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-            android:id="@+id/nav_movie_list"
-            android:icon="@drawable/ic_movies"
-            android:title="@string/movies"/>
+            android:id="@+id/nav_home"
+            android:icon="@drawable/selector_navigation_icon_home"
+            android:title="@string/nav_home"/>
 
     <item
-            android:id="@+id/nav_pref"
-            android:icon="@drawable/ic_movies"
-            android:title="@string/movies"/>
+            android:id="@+id/nav_movie_list"
+            android:icon="@drawable/selector_navigation_icon_movie"
+            android:title="@string/nav_movie"/>
 
 </menu>
