+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">Top Movies</string>
-    <string name="movies">Movies</string>
+    <string name="nav_home">Home</string>
+    <string name="nav_movie">Movies</string>
 </resources>
